requirements_docker.txt
numpy==1.20.0
opencv-python==4.6.0.66
pandas==1.3.0
matplotlib==3.4.2
boto3
scipy==1.5.4
scikit-learn
seaborn
tqdm==4.64.0
protobuf==3.20.0
albumentations==1.3.0
torch-summary==1.4.5